Struct vrd::Random

#[non_exhaustive]
pub struct Random {
    pub mt: [u32; 624],
    pub mti: usize,
}

The Random struct is used to generate random numbers using the Mersenne Twister algorithm.

It contains an array of unsigned 32-bit integers and an index used to generate random numbers. The array contains 624 elements and the index is used to generate random numbers from the array.

The index is incremented after each random number is generated. When the index reaches 624, the array is reinitialized and the index is reset to 0.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

mt: [u32; 624]

The array of unsigned 32-bit integers used to generate random numbers

mti: usize

The current index of the array used in the generation of random numbers

Implementations

impl Random

pub fn bool(&mut self, probability: f64) -> bool

Returns a random bool with a probability that can be set.

pub fn bytes(&mut self, len: usize) -> Vec<u8>

Returns a vector of random bytes of the given length

pub fn char(&mut self) -> char

Returns a random char within the range ‘a’..=‘z’

pub fn choose<'a, T>(&'a mut self, values: &'a [T]) -> Option<&T>

Returns a random element from a slice of values

pub fn float(&mut self) -> f32

Returns a random float.

pub fn int(&mut self, min: i32, max: i32) -> i32

Returns a random integer within the given range

pub fn uint(&mut self, min: u32, max: u32) -> u32

Returns a random unsigned integer within the given range

pub fn double(&mut self) -> f64

Returns a random double.

pub fn new() -> Self

Returns new random number generator

pub fn pseudo(&mut self) -> u32

Returns pseudo random number

pub fn rand(&mut self) -> u32

Returns a random 32-bit unsigned integer.

pub fn random_range(&mut self, min: u32, max: u32) -> u32

Returns a random 32-bit unsigned integer within a given range.

pub fn range(&mut self, min: i32, max: i32) -> i32

Returns a random number within a given range.

pub fn seed(&mut self, seed: u32)

Seeds the random number generator with a given value.

pub fn twist(&mut self)

Twists the state of the random number generator.