Module vrd::macros

The macros module contains functions for generating macros.

Macros for the Random (VRD) crate.

This module contains macros that simplify working with the Random (VRD) crate.

The macros included in this module allow for quick and easy access to common functionality provided by the Random (VRD) crate.

These macros can greatly simplify code that uses the Random (VRD) crate, making it easier to read and maintain.

Macros

The following macros are provided by this module:

rand_bool!() - Generate a random boolean with the provided probability

rand_bytes!() - Generate a vector of random bytes with the provided length

rand_char!() - Generate a random char within the range ‘a’..=‘z’

rand_choose!() - Generate a random element from a slice of values

rand_float!() - Generate a random float

rand_int!() - Generate a random integer within the given range

rand_uint!() - Generate a random 32-bit unsigned integer within the given range

rand_double!() - Generate a random double

rand_new!() - Generate a new random number

rand_pseudo!() - Generate a pseudo random number

rand_range!() - Generate a random number within the given range

rand_seed!() - Seed the provided Random (VRD) struct with the given value

rand_twist!() - Twist the state of the provided Random (VRD) struct